instr/meter: remove old code, move metric package
