+++ pipe/pipe_test.go
@@ -180,89 +180,3 @@ func newPipe(t *testing.T) *pipe.Pipe {
 	assert.Nil(t, err)
 	return p
 }
-
-// func TestMetricsEmpty(t *testing.T) {
-// 	p, err := pipe.New(sampleRate)
-// 	assert.Nil(t, err)
-// 	mc := p.Measure()
-// 	assert.Nil(t, mc)
-// 	_ = pipe.Wait(p.Close())
-// }
-
-// func TestMetricsBadID(t *testing.T) {
-// 	proc := &mock.Processor{UID: phono.NewUID()}
-// 	p, err := pipe.New(sampleRate, pipe.WithProcessors(proc))
-// 	assert.Nil(t, err)
-// 	mc := p.Measure(proc.ID() + "bad")
-// 	assert.Nil(t, mc)
-// 	_ = pipe.Wait(p.Close())
-// }
-
-// func TestMetrics(t *testing.T) {
-// 	pump := &mock.Pump{
-// 		UID:         phono.NewUID(),
-// 		Limit:       measureTests.Limit,
-// 		Interval:    measureTests.interval,
-// 		BufferSize:  measureTests.BufferSize,
-// 		NumChannels: measureTests.NumChannels,
-// 	}
-// 	proc := &mock.Processor{UID: phono.NewUID()}
-// 	sink := &mock.Sink{UID: phono.NewUID()}
-// 	p, err := pipe.New(
-// 		sampleRate,
-// 		pipe.WithName("Test Metrics"),
-// 		pipe.WithPump(pump),
-// 		pipe.WithProcessors(proc),
-// 		pipe.WithSinks(sink),
-// 	)
-// 	assert.Nil(t, err)
-
-// 	var mc <-chan pipe.Measure
-
-// 	// zero measures
-// 	mc = p.Measure()
-// 	assert.NotNil(t, mc)
-// 	for m := range mc {
-// 		assert.NotNil(t, m)
-// 		assert.Equal(t, time.Time{}, m.Start)
-// 		assert.Equal(t, time.Duration(0), m.Elapsed)
-// 		for _, c := range m.Counters {
-// 			assert.Equal(t, int64(0), c.Messages())
-// 			assert.Equal(t, int64(0), c.Samples())
-// 			assert.Equal(t, time.Duration(0), c.Duration())
-// 		}
-// 		switch m.ID {
-// 		case pump.ID():
-// 		case proc.ID():
-// 		case sink.ID():
-// 		default:
-// 			t.Errorf("Measure with empty id")
-// 		}
-// 	}
-
-// 	start := time.Now()
-// 	p.Run()
-// 	err = pipe.Wait(p.Pause())
-// 	assert.Nil(t, err)
-// 	mc = p.Measure(pump.ID(), proc.ID(), sink.ID())
-// 	// measure diring pausing
-// 	for m := range mc {
-// 		assert.NotNil(t, m)
-// 		assert.True(t, start.Before(m.Start))
-// 		assert.True(t, measureTests.interval < m.Elapsed)
-// 		switch m.ID {
-// 		case pump.ID():
-// 		case proc.ID():
-// 		case sink.ID():
-// 		default:
-// 			t.Errorf("Measure with empty id")
-// 		}
-// 	}
-// 	err = pipe.Wait(p.Resume())
-// 	assert.Nil(t, err)
-
-// 	p.Run()
-// 	mc = p.Measure()
-// 	_ = pipe.Wait(p.Close())
-// 	goleak.VerifyNoLeaks(t)
-// }
